Regression Machine Learning Model
This is my first project based on Machine Learning, I trained a regression model on the bike count dataset. The goal of this dataset is to predict the target variable, which is the Rented Bike Count. The features used in the dataset include:

1. Date
2. Hour
3. Temperature (°C)
4. Humidity (%)
5. Wind Speed (m/s)
6. Visibility (10m)
7. Dew Point Temperature
8. Solar Radiation (MJ/m²)
9. Rainfall (mm)
10. Snowfall (cm)
11. Seasons
12. Holiday
13. Functioning Day

Here is the regression model analysis:

1. XGBoost Regressor: 0.9247
2. Random Forest Regressor: 0.9101
3. Gradient Boosting Regressor: 0.9031
4. Support Vector Regressor: 0.8838
5. Decision Tree Regressor: 0.8391
6. Linear Regression: 0.7116

XGBoost Regressor, achieved the highest R² score so it will perform better on this data set